Add files via upload
Figured out curvature centerpoints for incomplete cylinders like fillets with the following approach: when we remove faces for cylindrical approximation, hang onto the normal directions. Project a line in that direction from the centerpoint of that face. The point where each of those lines intersect is the centerpoint.

Our next challenge is to set up the starting and ending points for those partial curves based on the interesection points with the faces we havent removed. Should be fun!